# lib/api-zod/src/generated/types/accessibilityProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CrisisIQ API
 * OpenAPI spec version: 0.1.0
 */

export interface AccessibilityProfile {
  deaf: boolean;
  wheelchair: boolean;
  visuallyImpaired: boolean;
  nonVerbal: boolean;
  cognitive: boolean;
}