# lib/api-zod/src/generated/types/coachMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CrisisIQ API
 * OpenAPI spec version: 0.1.0
 */
import type { CoachMessageRole } from "./coachMessageRole";

export interface CoachMessage {
  role: CoachMessageRole;
  text: string;
  timestamp: string;
}