# lib/api-zod/src/generated/types/createGuestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CrisisIQ API
 * OpenAPI spec version: 0.1.0
 */
import type { AccessibilityProfile } from "./accessibilityProfile";

export interface CreateGuestBody {
  name: string;
  email?: string | null;
  phone?: string | null;
  room: string;
  language: string;
  accessibility: AccessibilityProfile;
}