# lib/api-zod/src/generated/types/createIncidentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CrisisIQ API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateIncidentBodyType } from "./createIncidentBodyType";

export interface CreateIncidentBody {
  type: CreateIncidentBodyType;
  room: string;
  floor: number;
  guestId?: string | null;
  description?: string | null;
}