# lib/api-zod/src/generated/types/debriefReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CrisisIQ API
 * OpenAPI spec version: 0.1.0
 */
import type { DebriefReportTimelineItem } from "./debriefReportTimelineItem";

export interface DebriefReport {
  incidentId: string;
  executiveSummary: string;
  timeline: DebriefReportTimelineItem[];
  responseTime: string;
  whatWorked: string[];
  whatFailed: string[];
  recommendations: string[];
}