# lib/api-zod/src/generated/types/socialRadar.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CrisisIQ API
 * OpenAPI spec version: 0.1.0
 */
import type { SocialRadarPostsItem } from "./socialRadarPostsItem";
import type { SocialRadarThreatLevel } from "./socialRadarThreatLevel";

export interface SocialRadar {
  threatLevel: SocialRadarThreatLevel;
  summary: string;
  posts: SocialRadarPostsItem[];
}