# lib/api-zod/src/generated/types/surveySubmission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CrisisIQ API
 * OpenAPI spec version: 0.1.0
 */

export interface SurveySubmission {
  guestId: string;
  answers: number[];
}